requirements.txt
Flask==2.3.2
openai-whisper==20230314
pydub==0.25.1
torch==2.0.1
numpy==1.24.3
ffmpeg-python==0.2.0